Update Rust to the latest nightly

+++ crates/pallet-grandpa-finality-verifier/src/tests/justification.rs
@@ -100,7 +100,7 @@ pub(crate) fn make_justification_for_header<H: HeaderT>(
         };
 
         // Note: Adding 1 to account for the target header
-        let chain = generate_chain(i as u32, depth + 1, &header);
+        let chain = generate_chain(i, depth + 1, &header);
 
         // We don't include our finality target header in the vote ancestries
         for child in &chain[1..] {
